multiple project pickers added and routed
